Move module converters to internal/model/idl
BufferTypeConverter, RawJsonTypeConverter, and FastestTypeConverter
follow the model/idl package convention persistence established at
api/model/idl, mirrored on the internal side.
