+++ kinotic-core/src/main/java/org/kinotic/core/internal/model/idl/BufferTypeConverter.java
@@ -1,22 +1,23 @@
-package org.kinotic.domain.internal.serializer;
+package org.kinotic.core.internal.model.idl;
 
-import org.kinotic.domain.api.model.RawJson;
+import io.vertx.core.buffer.Buffer;
 import org.kinotic.idl.api.directory.ConversionContext;
 import org.kinotic.idl.api.directory.SpecificTypeConverter;
-import org.kinotic.idl.api.schema.AnyC3Type;
+import org.kinotic.idl.api.schema.ArrayC3Type;
+import org.kinotic.idl.api.schema.ByteC3Type;
 import org.kinotic.idl.api.schema.C3Type;
 import org.springframework.core.ResolvableType;
 import org.springframework.stereotype.Component;
 
 /**
- * Converts a {@link RawJson} to an {@link AnyC3Type}: RawJsonSerializer writes its data as a raw JSON value,
- * so on the wire it is whatever JSON it holds.
+ * Converts a Vert.x {@link Buffer} to an {@link ArrayC3Type} of bytes: BinaryReturnValueConverter sends it
+ * straight through as raw bytes on the wire.
  * Created by Navíd Mitchell 🤪 on 7/22/26
  */
 @Component
-public class RawJsonTypeConverter implements SpecificTypeConverter {
+public class BufferTypeConverter implements SpecificTypeConverter {
 
-    private static final Class<?>[] supports = {RawJson.class};
+    private static final Class<?>[] supports = {Buffer.class};
 
     @Override
     public Class<?>[] supports() {
@@ -26,6 +27,6 @@ public class RawJsonTypeConverter implements SpecificTypeConverter {
     @Override
     public C3Type convert(ResolvableType resolvableType,
                           ConversionContext conversionContext) {
-        return new AnyC3Type();
+        return new ArrayC3Type().setContains(new ByteC3Type());
     }
 }
