+++ kinotic-persistence/src/main/java/org/kinotic/persistence/internal/model/idl/FastestTypeConverter.java
@@ -1,23 +1,22 @@
-package org.kinotic.core.internal.api.directory;
+package org.kinotic.persistence.internal.model.idl;
 
-import io.vertx.core.buffer.Buffer;
 import org.kinotic.idl.api.directory.ConversionContext;
 import org.kinotic.idl.api.directory.SpecificTypeConverter;
-import org.kinotic.idl.api.schema.ArrayC3Type;
-import org.kinotic.idl.api.schema.ByteC3Type;
+import org.kinotic.idl.api.schema.AnyC3Type;
 import org.kinotic.idl.api.schema.C3Type;
+import org.kinotic.persistence.api.model.FastestType;
 import org.springframework.core.ResolvableType;
 import org.springframework.stereotype.Component;
 
 /**
- * Converts a Vert.x {@link Buffer} to an {@link ArrayC3Type} of bytes: BinaryReturnValueConverter sends it
- * straight through as raw bytes on the wire.
+ * Converts a {@link FastestType} to an {@link AnyC3Type}: FastestTypeSerializer writes its wrapped data
+ * directly, so on the wire it is whatever JSON that data serializes to.
  * Created by Navíd Mitchell 🤪 on 7/22/26
  */
 @Component
-public class BufferTypeConverter implements SpecificTypeConverter {
+public class FastestTypeConverter implements SpecificTypeConverter {
 
-    private static final Class<?>[] supports = {Buffer.class};
+    private static final Class<?>[] supports = {FastestType.class};
 
     @Override
     public Class<?>[] supports() {
@@ -27,6 +26,6 @@ public class BufferTypeConverter implements SpecificTypeConverter {
     @Override
     public C3Type convert(ResolvableType resolvableType,
                           ConversionContext conversionContext) {
-        return new ArrayC3Type().setContains(new ByteC3Type());
+        return new AnyC3Type();
     }
 }
